support Travis CI

+++ src/test_create_reader.py
@@ -7,10 +7,12 @@ from pyzxing.reader import jar_filename, jar_path
 
 class TestCreateBarCodeReader(unittest.TestCase):
     def test_create_reader_no_local(self):
-        os.remove(jar_path+jar_filename)
+        if os.path.exists(jar_path+jar_filename):
+            os.remove(jar_path+jar_filename)
         self.reader = BarCodeReader()
 
     def test_create_reader_with_local(self):
-        shutil.rmtree(jar_path)
+        if os.path.exists(jar_path):
+            shutil.rmtree(jar_path)
         self.reader = BarCodeReader()
         self.reader = BarCodeReader()
